+++ src/commands/Command.ts
@@ -2,6 +2,7 @@
 export enum Command {
   init = "init",
   publish = "publish",
+  workflow = "workflow",
   version = "version",
   setup = "setup",
   cleanup = "cleanup",